+++ lib/src/xxhash64.dart
@@ -7,17 +7,19 @@ import 'package:hashlib/src/algorithms/xxhash64.dart';
 import 'package:hashlib/src/core/block_hash.dart';
 
 /// An instance of [XXHash64] with seed = 0
-const XXHash64 xxh64 = XXHash64(0);
+const XXHash64 xxhash64 = XXHash64(0);
 
-/// This is an implementation of 64-bit Hash algorithm from xxHash family that
-/// is derived from https://github.com/Cyan4973/xxHash
+/// XXHash64 is a fast and efficient non-cryptographic hash function for
+/// 64-bit platforms. It is designed for producing a quick and reliable hash
+/// value for a given data, which can be used for many applications, such as
+/// checksum, data validation, etc. In addition, it has a good distribution of
+/// hash values, which helps to reduce [collisions][wiki].
 ///
-/// It has been tested with [Austin Appleby's excellent SMHasher][smhasher]
-/// test suite, and passes all tests, ensuring reasonable quality levels.
-/// A more details collision-ratio comparison can be found in [the wiki][wiki].
+/// This implementation was derived from https://github.com/Cyan4973/xxHash
 ///
-/// [smhasher]: https://github.com/aappleby/smhasher
 /// [wiki]: https://github.com/Cyan4973/xxHash/wiki/Collision-ratio-comparison
+///
+/// **WARNING: It should not be used for cryptographic purposes.**
 class XXHash64 extends BlockHashBase {
   final int seed;
 
@@ -30,22 +32,22 @@ class XXHash64 extends BlockHashBase {
   XXHash64 withSeed(int seed) => XXHash64(seed);
 }
 
-/// Gets the xxHash-64 remainder value of a String
+/// Gets the 64-bit xxHash value of a String
 ///
 /// Parameters:
 /// - [input] is the string to hash
 /// - The [encoding] is the encoding to use. Default is `input.codeUnits`
-int xxh64code(String input, [Encoding? encoding]) {
-  return xxh64.string(input, encoding).remainder();
+int xxhash64code(String input, [Encoding? encoding]) {
+  return xxhash64.string(input, encoding).remainder();
 }
 
-/// Extension to [String] to generate [xxh64] code
+/// Extension to [String] to generate [xxhash64] code
 extension XXHash64StringExtension on String {
-  /// Gets the xxHash-64 remainder value of a String
+  /// Gets the 64-bit xxHash value of a String
   ///
   /// Parameters:
   /// - If no [encoding] is defined, the `codeUnits` is used to get the bytes.
-  int xxh64code([Encoding? encoding]) {
-    return xxh64.string(this, encoding).remainder();
+  int xxhash64code([Encoding? encoding]) {
+    return xxhash64.string(this, encoding).remainder();
   }
 }