+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 - Adds xxHash64
   - New class: `XXHash64`
-  - New constants: `xxh64`, `xxh64code`
+  - New constants: `xxhash64`, `xxhash64code`
+  - String extension: `xxhash64code`
 - Adds xxHash32
   - New class: `XXHash32`
-  - New constants: `xxh32`, `xxh32code`
+  - New constants: `xxhash32`, `xxhash32code`
+  - String extension: `xxhash32code`
+- Adds XXH3
+  - New class: `XXH3`
+  - New constants: `xxh3`, `xxh3code`
+  - String extension: `xxh3code`
 - Internal changes:
   - Removes the parameters from `$finalize` method in `BlockHash`
   - Uses `>>>` instead of `>>`